deser: Separate macros tests from other tests
