+++ scylla-cql/src/deserialize/row_tests.rs
@@ -2,7 +2,7 @@ use assert_matches::assert_matches;
 use bytes::Bytes;
 
 use crate::deserialize::row::BuiltinDeserializationErrorKind;
-use crate::deserialize::{DeserializationError, FrameSlice, TypeCheckError, value};
+use crate::deserialize::{DeserializationError, FrameSlice, TypeCheckError};
 use crate::frame::response::result::{ColumnSpec, ColumnType, NativeType, TableSpec};
 
 use super::super::tests::{serialize_cells, spec};
@@ -92,12 +92,12 @@ fn val_str(s: &str) -> Option<Vec<u8>> {
 /// Wrapper for type check vs. deserialization errors, since we no longer convert
 /// TypeCheckError into DeserializationError.
 #[derive(Debug)]
-pub(crate) enum TestDeserializeError {
+enum TestDeserializeError {
     TypeCheck(TypeCheckError),
     Deserialization(DeserializationError),
 }
 
-pub(crate) fn deserialize<'frame, 'metadata, R>(
+fn deserialize<'frame, 'metadata, R>(
     specs: &'metadata [ColumnSpec<'metadata>],
     byts: &'frame Bytes,
 ) -> Result<R, TestDeserializeError>
@@ -113,7 +113,7 @@ where
 }
 
 #[track_caller]
-pub(crate) fn get_typck_err_inner(err: &TypeCheckError) -> &BuiltinTypeCheckError {
+fn get_typck_err_inner(err: &TypeCheckError) -> &BuiltinTypeCheckError {
     match err.downcast_ref() {
         Some(err) => err,
         None => panic!("not a BuiltinTypeCheckError: {err:?}"),
@@ -287,470 +287,6 @@ fn test_row_errors() {
             panic!("unexpected error kind: {}", err.kind)
         };
         assert_eq!(column_name, col_name);
-    }
-}
-
-fn specs_to_types<'a>(specs: &[ColumnSpec<'a>]) -> Vec<ColumnType<'a>> {
-    specs.iter().map(|spec| spec.typ().clone()).collect()
-}
-
-#[test]
-fn test_struct_deserialization_errors() {
-    // Loose ordering
-    {
-        #[derive(scylla_macros::DeserializeRow, PartialEq, Eq, Debug)]
-        #[scylla(crate = "crate")]
-        struct MyRow<'a> {
-            a: &'a str,
-            #[scylla(skip)]
-            x: String,
-            b: Option<i32>,
-            #[scylla(rename = "c")]
-            d: bool,
-            #[scylla(default_when_null)]
-            e: Option<i32>,
-        }
-
-        // Type check errors
-        {
-            // Missing column
-            {
-                let specs = [
-                    spec("a", ColumnType::Native(NativeType::Ascii)),
-                    spec("b", ColumnType::Native(NativeType::Int)),
-                ];
-                let err = MyRow::type_check(&specs).unwrap_err();
-                let err = get_typck_err_inner(&err);
-                assert_eq!(err.rust_name, std::any::type_name::<MyRow>());
-                assert_eq!(err.cql_types, specs_to_types(&specs));
-                let BuiltinTypeCheckErrorKind::ValuesMissingForColumns {
-                    column_names: ref missing_fields,
-                } = err.kind
-                else {
-                    panic!("unexpected error kind: {:?}", err.kind)
-                };
-                assert_eq!(missing_fields.as_slice(), &["c", "e"]);
-            }
-
-            // Duplicated column
-            {
-                let specs = [
-                    spec("a", ColumnType::Native(NativeType::Ascii)),
-                    spec("b", ColumnType::Native(NativeType::Int)),
-                    spec("a", ColumnType::Native(NativeType::Ascii)),
-                ];
-
-                let err = MyRow::type_check(&specs).unwrap_err();
-                let err = get_typck_err_inner(&err);
-                assert_eq!(err.rust_name, std::any::type_name::<MyRow>());
-                assert_eq!(err.cql_types, specs_to_types(&specs));
-                let BuiltinTypeCheckErrorKind::DuplicatedColumn {
-                    column_index,
-                    column_name,
-                } = err.kind
-                else {
-                    panic!("unexpected error kind: {:?}", err.kind)
-                };
-                assert_eq!(column_index, 2);
-                assert_eq!(column_name, "a");
-            }
-
-            // Unknown column
-            {
-                let specs = [
-                    spec("d", ColumnType::Native(NativeType::Counter)),
-                    spec("a", ColumnType::Native(NativeType::Ascii)),
-                    spec("b", ColumnType::Native(NativeType::Int)),
-                ];
-
-                let err = MyRow::type_check(&specs).unwrap_err();
-                let err = get_typck_err_inner(&err);
-                assert_eq!(err.rust_name, std::any::type_name::<MyRow>());
-                assert_eq!(err.cql_types, specs_to_types(&specs));
-                let BuiltinTypeCheckErrorKind::ColumnWithUnknownName {
-                    column_index,
-                    ref column_name,
-                } = err.kind
-                else {
-                    panic!("unexpected error kind: {:?}", err.kind)
-                };
-                assert_eq!(column_index, 0);
-                assert_eq!(column_name.as_str(), "d");
-            }
-
-            // Column incompatible types - column type check failed
-            {
-                let specs = [
-                    spec("b", ColumnType::Native(NativeType::Int)),
-                    spec("a", ColumnType::Native(NativeType::Blob)),
-                ];
-                let err = MyRow::type_check(&specs).unwrap_err();
-                let err = get_typck_err_inner(&err);
-                assert_eq!(err.rust_name, std::any::type_name::<MyRow>());
-                assert_eq!(err.cql_types, specs_to_types(&specs));
-                let BuiltinTypeCheckErrorKind::ColumnTypeCheckFailed {
-                    column_index,
-                    ref column_name,
-                    ref err,
-                } = err.kind
-                else {
-                    panic!("unexpected error kind: {:?}", err.kind)
-                };
-                assert_eq!(column_index, 1);
-                assert_eq!(column_name.as_str(), "a");
-                let err = value::tests::get_typeck_err_inner(err);
-                assert_eq!(err.rust_name, std::any::type_name::<&str>());
-                assert_eq!(err.cql_type, ColumnType::Native(NativeType::Blob));
-                assert_matches!(
-                    err.kind,
-                    value::BuiltinTypeCheckErrorKind::MismatchedType {
-                        expected: &[
-                            ColumnType::Native(NativeType::Ascii),
-                            ColumnType::Native(NativeType::Text)
-                        ]
-                    }
-                );
-            }
-        }
-
-        // Deserialization errors
-        {
-            // Got null
-            {
-                let specs = [
-                    spec("c", ColumnType::Native(NativeType::Boolean)),
-                    spec("a", ColumnType::Native(NativeType::Blob)),
-                    spec("b", ColumnType::Native(NativeType::Int)),
-                    spec("e", ColumnType::Native(NativeType::Int)),
-                ];
-
-                let err = MyRow::deserialize(ColumnIterator::new(
-                    &specs,
-                    FrameSlice::new(&serialize_cells([Some([true as u8])])),
-                ))
-                .unwrap_err();
-                let err = get_deser_err_inner(&err);
-                assert_eq!(err.rust_name, std::any::type_name::<MyRow>());
-                let BuiltinDeserializationErrorKind::RawColumnDeserializationFailed {
-                    column_index,
-                    ref column_name,
-                    ..
-                } = err.kind
-                else {
-                    panic!("unexpected error kind: {:?}", err.kind)
-                };
-                assert_eq!(column_index, 1);
-                assert_eq!(column_name, "a");
-            }
-
-            // Column deserialization failed
-            {
-                let specs = [
-                    spec("b", ColumnType::Native(NativeType::Int)),
-                    spec("a", ColumnType::Native(NativeType::Ascii)),
-                    spec("c", ColumnType::Native(NativeType::Boolean)),
-                    spec("e", ColumnType::Native(NativeType::Int)),
-                ];
-
-                let row_bytes = serialize_cells(
-                    [
-                        &0_i32.to_be_bytes(),
-                        "alamakota".as_bytes(),
-                        &42_i16.to_be_bytes(),
-                        &13_i32.to_be_bytes(),
-                    ]
-                    .map(Some),
-                );
-
-                let err = deserialize::<MyRow>(&specs, &row_bytes).unwrap_err();
-
-                let err = get_deser_err(&err);
-                assert_eq!(err.rust_name, std::any::type_name::<MyRow>());
-                let BuiltinDeserializationErrorKind::ColumnDeserializationFailed {
-                    column_index,
-                    ref column_name,
-                    ref err,
-                } = err.kind
-                else {
-                    panic!("unexpected error kind: {:?}", err.kind)
-                };
-                assert_eq!(column_index, 2);
-                assert_eq!(column_name.as_str(), "c");
-                let err = value::tests::get_deser_err_inner(err);
-                assert_eq!(err.rust_name, std::any::type_name::<bool>());
-                assert_eq!(err.cql_type, ColumnType::Native(NativeType::Boolean));
-                assert_matches!(
-                    err.kind,
-                    value::BuiltinDeserializationErrorKind::ByteLengthMismatch {
-                        expected: 1,
-                        got: 2,
-                    }
-                );
-            }
-        }
-    }
-
-    // Strict ordering
-    {
-        #[derive(scylla_macros::DeserializeRow, PartialEq, Eq, Debug)]
-        #[scylla(crate = "crate", flavor = "enforce_order")]
-        struct MyRow<'a> {
-            a: &'a str,
-            #[scylla(skip)]
-            x: String,
-            b: Option<i32>,
-            c: bool,
-            #[scylla(default_when_null)]
-            d: Option<i32>,
-        }
-
-        // Type check errors
-        {
-            // Too few columns
-            {
-                let specs = [spec("a", ColumnType::Native(NativeType::Text))];
-                let err = MyRow::type_check(&specs).unwrap_err();
-                let err = get_typck_err_inner(&err);
-                assert_eq!(err.rust_name, std::any::type_name::<MyRow>());
-                assert_eq!(err.cql_types, specs_to_types(&specs));
-                let BuiltinTypeCheckErrorKind::WrongColumnCount {
-                    rust_cols,
-                    cql_cols,
-                } = err.kind
-                else {
-                    panic!("unexpected error kind: {:?}", err.kind)
-                };
-                assert_eq!(rust_cols, 4);
-                assert_eq!(cql_cols, 1);
-            }
-
-            // Excess columns
-            {
-                let specs = [
-                    spec("a", ColumnType::Native(NativeType::Text)),
-                    spec("b", ColumnType::Native(NativeType::Int)),
-                    spec("c", ColumnType::Native(NativeType::Boolean)),
-                    spec("d", ColumnType::Native(NativeType::Int)),
-                    spec("e", ColumnType::Native(NativeType::Counter)),
-                ];
-                let err = MyRow::type_check(&specs).unwrap_err();
-                let err = get_typck_err_inner(&err);
-                assert_eq!(err.rust_name, std::any::type_name::<MyRow>());
-                assert_eq!(err.cql_types, specs_to_types(&specs));
-                let BuiltinTypeCheckErrorKind::WrongColumnCount {
-                    rust_cols,
-                    cql_cols,
-                } = err.kind
-                else {
-                    panic!("unexpected error kind: {:?}", err.kind)
-                };
-                assert_eq!(rust_cols, 4);
-                assert_eq!(cql_cols, 5);
-            }
-
-            // Renamed column name mismatch
-            {
-                let specs = [
-                    spec("a", ColumnType::Native(NativeType::Text)),
-                    spec("b", ColumnType::Native(NativeType::Int)),
-                    spec("e", ColumnType::Native(NativeType::Boolean)),
-                    spec("d", ColumnType::Native(NativeType::Int)),
-                ];
-                let err = MyRow::type_check(&specs).unwrap_err();
-                let err = get_typck_err_inner(&err);
-                assert_eq!(err.rust_name, std::any::type_name::<MyRow>());
-                let BuiltinTypeCheckErrorKind::ColumnNameMismatch {
-                    field_index,
-                    column_index,
-                    rust_column_name,
-                    ref db_column_name,
-                } = err.kind
-                else {
-                    panic!("unexpected error kind: {:?}", err.kind)
-                };
-                assert_eq!(field_index, 3);
-                assert_eq!(rust_column_name, "c");
-                assert_eq!(column_index, 2);
-                assert_eq!(db_column_name.as_str(), "e");
-            }
-
-            // Columns switched - column name mismatch
-            {
-                let specs = [
-                    spec("b", ColumnType::Native(NativeType::Int)),
-                    spec("a", ColumnType::Native(NativeType::Text)),
-                    spec("c", ColumnType::Native(NativeType::Boolean)),
-                    spec("d", ColumnType::Native(NativeType::Int)),
-                ];
-                let err = MyRow::type_check(&specs).unwrap_err();
-                let err = get_typck_err_inner(&err);
-                assert_eq!(err.rust_name, std::any::type_name::<MyRow>());
-                assert_eq!(err.cql_types, specs_to_types(&specs));
-                let BuiltinTypeCheckErrorKind::ColumnNameMismatch {
-                    field_index,
-                    column_index,
-                    rust_column_name,
-                    ref db_column_name,
-                } = err.kind
-                else {
-                    panic!("unexpected error kind: {:?}", err.kind)
-                };
-                assert_eq!(field_index, 0);
-                assert_eq!(column_index, 0);
-                assert_eq!(rust_column_name, "a");
-                assert_eq!(db_column_name.as_str(), "b");
-            }
-
-            // Column incompatible types - column type check failed
-            {
-                let specs = [
-                    spec("a", ColumnType::Native(NativeType::Blob)),
-                    spec("b", ColumnType::Native(NativeType::Int)),
-                    spec("c", ColumnType::Native(NativeType::Boolean)),
-                    spec("d", ColumnType::Native(NativeType::Int)),
-                ];
-                let err = MyRow::type_check(&specs).unwrap_err();
-                let err = get_typck_err_inner(&err);
-                assert_eq!(err.rust_name, std::any::type_name::<MyRow>());
-                assert_eq!(err.cql_types, specs_to_types(&specs));
-                let BuiltinTypeCheckErrorKind::ColumnTypeCheckFailed {
-                    column_index,
-                    ref column_name,
-                    ref err,
-                } = err.kind
-                else {
-                    panic!("unexpected error kind: {:?}", err.kind)
-                };
-                assert_eq!(column_index, 0);
-                assert_eq!(column_name.as_str(), "a");
-                let err = value::tests::get_typeck_err_inner(err);
-                assert_eq!(err.rust_name, std::any::type_name::<&str>());
-                assert_eq!(err.cql_type, ColumnType::Native(NativeType::Blob));
-                assert_matches!(
-                    err.kind,
-                    value::BuiltinTypeCheckErrorKind::MismatchedType {
-                        expected: &[
-                            ColumnType::Native(NativeType::Ascii),
-                            ColumnType::Native(NativeType::Text)
-                        ]
-                    }
-                );
-            }
-        }
-
-        // Deserialization errors
-        {
-            // Too few columns
-            {
-                let specs = [
-                    spec("a", ColumnType::Native(NativeType::Text)),
-                    spec("b", ColumnType::Native(NativeType::Int)),
-                    spec("c", ColumnType::Native(NativeType::Boolean)),
-                    spec("d", ColumnType::Native(NativeType::Int)),
-                ];
-
-                let err = MyRow::deserialize(ColumnIterator::new(
-                    &specs,
-                    FrameSlice::new(&serialize_cells([Some([true as u8])])),
-                ))
-                .unwrap_err();
-                let err = get_deser_err_inner(&err);
-                assert_eq!(err.rust_name, std::any::type_name::<MyRow>());
-                let BuiltinDeserializationErrorKind::RawColumnDeserializationFailed {
-                    column_index,
-                    ref column_name,
-                    ..
-                } = err.kind
-                else {
-                    panic!("unexpected error kind: {:?}", err.kind)
-                };
-                assert_eq!(column_index, 1);
-                assert_eq!(column_name, "b");
-            }
-
-            // Bad field format
-            {
-                let typ = [
-                    spec("a", ColumnType::Native(NativeType::Text)),
-                    spec("b", ColumnType::Native(NativeType::Int)),
-                    spec("c", ColumnType::Native(NativeType::Boolean)),
-                    spec("d", ColumnType::Native(NativeType::Int)),
-                ];
-
-                let row_bytes = serialize_cells(
-                    [
-                        (&b"alamakota"[..]),
-                        &42_i32.to_be_bytes(),
-                        &[true as u8],
-                        &13_i32.to_be_bytes(),
-                    ]
-                    .map(Some),
-                );
-
-                let row_bytes_too_short = row_bytes.slice(..row_bytes.len() - 1);
-                assert!(row_bytes.len() > row_bytes_too_short.len());
-
-                let err = deserialize::<MyRow>(&typ, &row_bytes_too_short).unwrap_err();
-
-                let err = get_deser_err(&err);
-                assert_eq!(err.rust_name, std::any::type_name::<MyRow>());
-                let BuiltinDeserializationErrorKind::RawColumnDeserializationFailed {
-                    column_index,
-                    ref column_name,
-                    ..
-                } = err.kind
-                else {
-                    panic!("unexpected error kind: {:?}", err.kind)
-                };
-                assert_eq!(column_index, 3);
-                assert_eq!(column_name, "d");
-            }
-
-            // Column deserialization failed
-            {
-                let specs = [
-                    spec("a", ColumnType::Native(NativeType::Text)),
-                    spec("b", ColumnType::Native(NativeType::Int)),
-                    spec("c", ColumnType::Native(NativeType::Boolean)),
-                    spec("d", ColumnType::Native(NativeType::Int)),
-                ];
-
-                let row_bytes = serialize_cells(
-                    [
-                        &b"alamakota"[..],
-                        &42_i64.to_be_bytes(),
-                        &[true as u8],
-                        &13_i32.to_be_bytes(),
-                    ]
-                    .map(Some),
-                );
-
-                let err = deserialize::<MyRow>(&specs, &row_bytes).unwrap_err();
-
-                let err = get_deser_err(&err);
-                assert_eq!(err.rust_name, std::any::type_name::<MyRow>());
-                let BuiltinDeserializationErrorKind::ColumnDeserializationFailed {
-                    column_index: field_index,
-                    ref column_name,
-                    ref err,
-                } = err.kind
-                else {
-                    panic!("unexpected error kind: {:?}", err.kind)
-                };
-                assert_eq!(column_name.as_str(), "b");
-                assert_eq!(field_index, 2);
-                let err = value::tests::get_deser_err_inner(err);
-                assert_eq!(err.rust_name, std::any::type_name::<Option<i32>>());
-                assert_eq!(err.cql_type, ColumnType::Native(NativeType::Int));
-                assert_matches!(
-                    err.kind,
-                    value::BuiltinDeserializationErrorKind::ByteLengthMismatch {
-                        expected: 4,
-                        got: 8,
-                    }
-                );
-            }
-        }
     }
 }
