+++ scylla-cql/src/deserialize/value_tests.rs
@@ -1,6 +1,5 @@
 use assert_matches::assert_matches;
 use bytes::{BufMut, Bytes, BytesMut};
-use scylla_macros::DeserializeValue;
 use uuid::Uuid;
 
 use std::borrow::Cow;
@@ -28,7 +27,7 @@ use super::{
     BuiltinDeserializationError, BuiltinDeserializationErrorKind, BuiltinTypeCheckError,
     BuiltinTypeCheckErrorKind, DeserializeValue, ListlikeIterator, MapDeserializationErrorKind,
     MapIterator, MapTypeCheckErrorKind, MaybeEmpty, SetOrListDeserializationErrorKind,
-    SetOrListTypeCheckErrorKind, UdtDeserializationErrorKind, UdtTypeCheckErrorKind, mk_deser_err,
+    SetOrListTypeCheckErrorKind, mk_deser_err,
 };
 
 #[test]
@@ -1262,469 +1261,6 @@ fn test_cow() {
     }
 }
 
-pub(crate) fn udt_def_with_fields(
-    fields: impl IntoIterator<Item = (impl Into<Cow<'static, str>>, ColumnType<'static>)>,
-) -> ColumnType<'static> {
-    ColumnType::UserDefinedType {
-        frozen: false,
-        definition: Arc::new(UserDefinedType {
-            name: "udt".into(),
-            keyspace: "ks".into(),
-            field_types: fields.into_iter().map(|(s, t)| (s.into(), t)).collect(),
-        }),
-    }
-}
-
-#[must_use]
-struct UdtSerializer {
-    buf: BytesMut,
-}
-
-impl UdtSerializer {
-    fn new() -> Self {
-        Self {
-            buf: BytesMut::default(),
-        }
-    }
-
-    fn field(mut self, field_bytes: &[u8]) -> Self {
-        append_bytes(&mut self.buf, field_bytes);
-        self
-    }
-
-    fn null_field(mut self) -> Self {
-        append_null(&mut self.buf);
-        self
-    }
-
-    fn finalize(&self) -> Bytes {
-        make_bytes(&self.buf)
-    }
-}
-
-// Do not remove. It's not used in tests but we keep it here to check that
-// we properly ignore warnings about unused variables, unnecessary `mut`s
-// etc. that usually pop up when generating code for empty structs.
-#[derive(scylla_macros::DeserializeValue)]
-#[scylla(crate = crate)]
-struct TestUdtWithNoFieldsUnordered {}
-
-#[derive(scylla_macros::DeserializeValue)]
-#[scylla(crate = crate, flavor = "enforce_order")]
-struct TestUdtWithNoFieldsOrdered {}
-
-// If deserialize is never called, rust warns that the struct is never constructed.
-// We don't want to `expect(dead_code)` on struct definitions, because that could silence
-// some warnings that this test is supposed to prevent.
-#[expect(unreachable_code, dead_code)]
-fn dummy_deserialize_udts() {
-    let _ = deserialize::<TestUdtWithNoFieldsUnordered>(todo!(), todo!()).unwrap();
-    let _ = deserialize::<TestUdtWithNoFieldsOrdered>(todo!(), todo!()).unwrap();
-}
-
-#[test]
-fn test_udt_loose_ordering() {
-    #[derive(scylla_macros::DeserializeValue, PartialEq, Eq, Debug)]
-    #[scylla(crate = "crate")]
-    struct Udt<'a> {
-        a: &'a str,
-        #[scylla(skip)]
-        x: String,
-        #[scylla(allow_missing)]
-        b: Option<i32>,
-        #[scylla(default_when_null)]
-        c: i64,
-    }
-
-    // UDT fields in correct same order.
-    {
-        let udt_bytes = UdtSerializer::new()
-            .field("The quick brown fox".as_bytes())
-            .field(&42_i32.to_be_bytes())
-            .field(&2137_i64.to_be_bytes())
-            .finalize();
-        let typ = udt_def_with_fields([
-            ("a", ColumnType::Native(NativeType::Text)),
-            ("b", ColumnType::Native(NativeType::Int)),
-            ("c", ColumnType::Native(NativeType::BigInt)),
-        ]);
-
-        let udt = deserialize::<Udt<'_>>(&typ, &udt_bytes).unwrap();
-        assert_eq!(
-            udt,
-            Udt {
-                a: "The quick brown fox",
-                x: String::new(),
-                b: Some(42),
-                c: 2137,
-            }
-        );
-    }
-
-    // The last two UDT field are missing in serialized form - it should treat it
-    // as if there were nulls at the end.
-    {
-        let udt_bytes = UdtSerializer::new()
-            .field("The quick brown fox".as_bytes())
-            .finalize();
-        let typ = udt_def_with_fields([
-            ("a", ColumnType::Native(NativeType::Text)),
-            ("b", ColumnType::Native(NativeType::Int)),
-            ("c", ColumnType::Native(NativeType::BigInt)),
-        ]);
-
-        let udt = deserialize::<Udt<'_>>(&typ, &udt_bytes).unwrap();
-        assert_eq!(
-            udt,
-            Udt {
-                a: "The quick brown fox",
-                x: String::new(),
-                b: None,
-                c: 0,
-            }
-        );
-    }
-
-    // UDT fields switched - should still work.
-    {
-        let udt_bytes = UdtSerializer::new()
-            .field(&42_i32.to_be_bytes())
-            .field("The quick brown fox".as_bytes())
-            .field(&2137_i64.to_be_bytes())
-            .finalize();
-        let typ = udt_def_with_fields([
-            ("b", ColumnType::Native(NativeType::Int)),
-            ("a", ColumnType::Native(NativeType::Text)),
-            ("c", ColumnType::Native(NativeType::BigInt)),
-        ]);
-
-        let udt = deserialize::<Udt<'_>>(&typ, &udt_bytes).unwrap();
-        assert_eq!(
-            udt,
-            Udt {
-                a: "The quick brown fox",
-                x: String::new(),
-                b: Some(42),
-                c: 2137,
-            }
-        );
-    }
-
-    // An excess UDT field - should still work.
-    {
-        let udt_bytes = UdtSerializer::new()
-            .field(&12_i8.to_be_bytes())
-            .field(&42_i32.to_be_bytes())
-            .field("The quick brown fox".as_bytes())
-            .field(&2137_i64.to_be_bytes())
-            .finalize();
-        let typ = udt_def_with_fields([
-            ("d", ColumnType::Native(NativeType::TinyInt)),
-            ("b", ColumnType::Native(NativeType::Int)),
-            ("a", ColumnType::Native(NativeType::Text)),
-            ("c", ColumnType::Native(NativeType::BigInt)),
-        ]);
-
-        Udt::type_check(&typ).unwrap();
-        let udt = deserialize::<Udt<'_>>(&typ, &udt_bytes).unwrap();
-        assert_eq!(
-            udt,
-            Udt {
-                a: "The quick brown fox",
-                x: String::new(),
-                b: Some(42),
-                c: 2137,
-            }
-        );
-    }
-
-    // Only field 'a' is present
-    {
-        let udt_bytes = UdtSerializer::new()
-            .field("The quick brown fox".as_bytes())
-            .finalize();
-        let typ = udt_def_with_fields([
-            ("a", ColumnType::Native(NativeType::Text)),
-            ("c", ColumnType::Native(NativeType::BigInt)),
-        ]);
-
-        let udt = deserialize::<Udt<'_>>(&typ, &udt_bytes).unwrap();
-        assert_eq!(
-            udt,
-            Udt {
-                a: "The quick brown fox",
-                x: String::new(),
-                b: None,
-                c: 0,
-            }
-        );
-    }
-
-    // Wrong column type
-    {
-        let typ = udt_def_with_fields([("a", ColumnType::Native(NativeType::Text))]);
-        Udt::type_check(&typ).unwrap_err();
-    }
-
-    // Missing required column
-    {
-        let typ = udt_def_with_fields([("b", ColumnType::Native(NativeType::Int))]);
-        Udt::type_check(&typ).unwrap_err();
-    }
-}
-
-#[test]
-fn test_udt_strict_ordering() {
-    #[derive(scylla_macros::DeserializeValue, PartialEq, Eq, Debug)]
-    #[scylla(crate = "crate", flavor = "enforce_order")]
-    struct Udt<'a> {
-        #[scylla(default_when_null)]
-        a: &'a str,
-        #[scylla(skip)]
-        x: String,
-        #[scylla(allow_missing)]
-        b: Option<i32>,
-    }
-
-    // UDT fields in correct same order
-    {
-        let udt_bytes = UdtSerializer::new()
-            .field("The quick brown fox".as_bytes())
-            .field(&42i32.to_be_bytes())
-            .finalize();
-        let typ = udt_def_with_fields([
-            ("a", ColumnType::Native(NativeType::Text)),
-            ("b", ColumnType::Native(NativeType::Int)),
-        ]);
-
-        let udt = deserialize::<Udt<'_>>(&typ, &udt_bytes).unwrap();
-        assert_eq!(
-            udt,
-            Udt {
-                a: "The quick brown fox",
-                x: String::new(),
-                b: Some(42),
-            }
-        );
-    }
-
-    // The last UDT field is missing in serialized form - it should treat
-    // as if there were null at the end
-    {
-        let udt_bytes = UdtSerializer::new()
-            .field("The quick brown fox".as_bytes())
-            .finalize();
-        let typ = udt_def_with_fields([
-            ("a", ColumnType::Native(NativeType::Text)),
-            ("b", ColumnType::Native(NativeType::Int)),
-        ]);
-
-        let udt = deserialize::<Udt<'_>>(&typ, &udt_bytes).unwrap();
-        assert_eq!(
-            udt,
-            Udt {
-                a: "The quick brown fox",
-                x: String::new(),
-                b: None,
-            }
-        );
-    }
-
-    // An excess field at the end of UDT
-    {
-        let udt_bytes = UdtSerializer::new()
-            .field("The quick brown fox".as_bytes())
-            .field(&42_i32.to_be_bytes())
-            .field(&(true as i8).to_be_bytes())
-            .finalize();
-        let typ = udt_def_with_fields([
-            ("a", ColumnType::Native(NativeType::Text)),
-            ("b", ColumnType::Native(NativeType::Int)),
-            ("d", ColumnType::Native(NativeType::Boolean)),
-        ]);
-        let udt = deserialize::<Udt<'_>>(&typ, &udt_bytes).unwrap();
-        assert_eq!(
-            udt,
-            Udt {
-                a: "The quick brown fox",
-                x: String::new(),
-                b: Some(42),
-            }
-        );
-    }
-
-    // An excess field at the end of UDT, when such are forbidden
-    {
-        #[derive(scylla_macros::DeserializeValue, PartialEq, Eq, Debug)]
-        #[scylla(crate = "crate", flavor = "enforce_order", forbid_excess_udt_fields)]
-        struct Udt<'a> {
-            a: &'a str,
-            #[scylla(skip)]
-            x: String,
-            b: Option<i32>,
-        }
-
-        let typ = udt_def_with_fields([
-            ("a", ColumnType::Native(NativeType::Text)),
-            ("b", ColumnType::Native(NativeType::Int)),
-            ("d", ColumnType::Native(NativeType::Boolean)),
-        ]);
-
-        Udt::type_check(&typ).unwrap_err();
-    }
-
-    // UDT fields switched - will not work
-    {
-        let typ = udt_def_with_fields([
-            ("b", ColumnType::Native(NativeType::Int)),
-            ("a", ColumnType::Native(NativeType::Text)),
-        ]);
-        Udt::type_check(&typ).unwrap_err();
-    }
-
-    // Wrong column type
-    {
-        let typ = udt_def_with_fields([
-            ("a", ColumnType::Native(NativeType::Int)),
-            ("b", ColumnType::Native(NativeType::Int)),
-        ]);
-        Udt::type_check(&typ).unwrap_err();
-    }
-
-    // Missing required column
-    {
-        let typ = udt_def_with_fields([("b", ColumnType::Native(NativeType::Int))]);
-        Udt::type_check(&typ).unwrap_err();
-    }
-
-    // Missing non-required column
-    {
-        let udt_bytes = UdtSerializer::new().field(b"kotmaale").finalize();
-        let typ = udt_def_with_fields([("a", ColumnType::Native(NativeType::Text))]);
-
-        let udt = deserialize::<Udt<'_>>(&typ, &udt_bytes).unwrap();
-        assert_eq!(
-            udt,
-            Udt {
-                a: "kotmaale",
-                x: String::new(),
-                b: None,
-            }
-        );
-    }
-
-    // The first field is null, but `default_when_null` prevents failure.
-    {
-        let udt_bytes = UdtSerializer::new()
-            .null_field()
-            .field(&42i32.to_be_bytes())
-            .finalize();
-        let typ = udt_def_with_fields([
-            ("a", ColumnType::Native(NativeType::Text)),
-            ("b", ColumnType::Native(NativeType::Int)),
-        ]);
-
-        let udt = deserialize::<Udt<'_>>(&typ, &udt_bytes).unwrap();
-        assert_eq!(
-            udt,
-            Udt {
-                a: "",
-                x: String::new(),
-                b: Some(42),
-            }
-        );
-    }
-}
-
-#[test]
-fn test_udt_no_name_check() {
-    #[derive(scylla_macros::DeserializeValue, PartialEq, Eq, Debug)]
-    #[scylla(crate = "crate", flavor = "enforce_order", skip_name_checks)]
-    struct Udt<'a> {
-        a: &'a str,
-        #[scylla(skip)]
-        x: String,
-        b: Option<i32>,
-    }
-
-    // UDT fields in correct same order
-    {
-        let udt_bytes = UdtSerializer::new()
-            .field("The quick brown fox".as_bytes())
-            .field(&42i32.to_be_bytes())
-            .finalize();
-        let typ = udt_def_with_fields([
-            ("a", ColumnType::Native(NativeType::Text)),
-            ("b", ColumnType::Native(NativeType::Int)),
-        ]);
-
-        let udt = deserialize::<Udt<'_>>(&typ, &udt_bytes).unwrap();
-        assert_eq!(
-            udt,
-            Udt {
-                a: "The quick brown fox",
-                x: String::new(),
-                b: Some(42),
-            }
-        );
-    }
-
-    // Correct order of UDT fields, but different names - should still succeed
-    {
-        let udt_bytes = UdtSerializer::new()
-            .field("The quick brown fox".as_bytes())
-            .field(&42i32.to_be_bytes())
-            .finalize();
-        let typ = udt_def_with_fields([
-            ("k", ColumnType::Native(NativeType::Text)),
-            ("l", ColumnType::Native(NativeType::Int)),
-        ]);
-
-        let udt = deserialize::<Udt<'_>>(&typ, &udt_bytes).unwrap();
-        assert_eq!(
-            udt,
-            Udt {
-                a: "The quick brown fox",
-                x: String::new(),
-                b: Some(42),
-            }
-        );
-    }
-}
-
-#[test]
-fn test_udt_cross_rename_fields() {
-    #[derive(scylla_macros::DeserializeValue, PartialEq, Eq, Debug)]
-    #[scylla(crate = crate)]
-    struct TestUdt {
-        #[scylla(rename = "b")]
-        a: i32,
-        #[scylla(rename = "a")]
-        b: String,
-    }
-
-    // UDT fields switched - should still work.
-    {
-        let udt_bytes = UdtSerializer::new()
-            .field("The quick brown fox".as_bytes())
-            .field(&42_i32.to_be_bytes())
-            .finalize();
-        let typ = udt_def_with_fields([
-            ("a", ColumnType::Native(NativeType::Text)),
-            ("b", ColumnType::Native(NativeType::Int)),
-        ]);
-
-        let udt = deserialize::<TestUdt>(&typ, &udt_bytes).unwrap();
-        assert_eq!(
-            udt,
-            TestUdt {
-                a: 42,
-                b: "The quick brown fox".to_owned(),
-            }
-        );
-    }
-}
-
 #[test]
 fn test_custom_type_parser() {
     #[derive(Default, Debug, PartialEq, Eq)]
@@ -1765,12 +1301,12 @@ fn test_custom_type_parser() {
 /// as separate variants, avoiding the need to wrap TypeCheckError into
 /// DeserializationError.
 #[derive(Debug)]
-pub(crate) enum TestDeserializeError {
+enum TestDeserializeError {
     TypeCheck(TypeCheckError),
     Deserialization(DeserializationError),
 }
 
-pub(crate) fn deserialize<'frame, 'metadata, T>(
+fn deserialize<'frame, 'metadata, T>(
     typ: &'metadata ColumnType<'metadata>,
     bytes: &'frame Bytes,
 ) -> Result<T, TestDeserializeError>
@@ -2761,411 +2297,6 @@ fn test_cow_errors() {
 }
 
 #[test]
-fn test_udt_errors() {
-    // Loose ordering
-    {
-        #[derive(scylla_macros::DeserializeValue, PartialEq, Eq, Debug)]
-        #[scylla(crate = "crate", forbid_excess_udt_fields)]
-        struct Udt<'a> {
-            a: &'a str,
-            #[scylla(skip)]
-            x: String,
-            #[scylla(allow_missing)]
-            b: Option<i32>,
-            #[scylla(default_when_null)]
-            c: bool,
-        }
-
-        // Type check errors
-        {
-            // Not UDT
-            {
-                let typ = ColumnType::Collection {
-                    frozen: false,
-                    typ: CollectionType::Map(
-                        Box::new(ColumnType::Native(NativeType::Ascii)),
-                        Box::new(ColumnType::Native(NativeType::Blob)),
-                    ),
-                };
-                let err = Udt::type_check(&typ).unwrap_err();
-                let err = get_typeck_err_inner(&err);
-                assert_eq!(err.rust_name, std::any::type_name::<Udt>());
-                assert_eq!(err.cql_type, typ);
-                let BuiltinTypeCheckErrorKind::UdtError(UdtTypeCheckErrorKind::NotUdt) = err.kind
-                else {
-                    panic!("unexpected error kind: {:?}", err.kind)
-                };
-            }
-
-            // UDT missing fields
-            {
-                let typ = udt_def_with_fields([("c", ColumnType::Native(NativeType::Boolean))]);
-                let err = Udt::type_check(&typ).unwrap_err();
-                let err = get_typeck_err_inner(&err);
-                assert_eq!(err.rust_name, std::any::type_name::<Udt>());
-                assert_eq!(err.cql_type, typ);
-                let BuiltinTypeCheckErrorKind::UdtError(
-                    UdtTypeCheckErrorKind::ValuesMissingForUdtFields {
-                        field_names: ref missing_fields,
-                    },
-                ) = err.kind
-                else {
-                    panic!("unexpected error kind: {:?}", err.kind)
-                };
-                assert_eq!(missing_fields.as_slice(), &["a"]);
-            }
-
-            // excess fields in UDT
-            {
-                let typ = udt_def_with_fields([
-                    ("d", ColumnType::Native(NativeType::Boolean)),
-                    ("a", ColumnType::Native(NativeType::Text)),
-                    ("b", ColumnType::Native(NativeType::Int)),
-                ]);
-                let err = Udt::type_check(&typ).unwrap_err();
-                let err = get_typeck_err_inner(&err);
-                assert_eq!(err.rust_name, std::any::type_name::<Udt>());
-                assert_eq!(err.cql_type, typ);
-                let BuiltinTypeCheckErrorKind::UdtError(UdtTypeCheckErrorKind::ExcessFieldInUdt {
-                    ref db_field_name,
-                }) = err.kind
-                else {
-                    panic!("unexpected error kind: {:?}", err.kind)
-                };
-                assert_eq!(db_field_name.as_str(), "d");
-            }
-
-            // missing UDT field
-            {
-                let typ = udt_def_with_fields([
-                    ("b", ColumnType::Native(NativeType::Int)),
-                    ("a", ColumnType::Native(NativeType::Text)),
-                ]);
-                let err = Udt::type_check(&typ).unwrap_err();
-                let err = get_typeck_err_inner(&err);
-                assert_eq!(err.rust_name, std::any::type_name::<Udt>());
-                assert_eq!(err.cql_type, typ);
-                let BuiltinTypeCheckErrorKind::UdtError(
-                    UdtTypeCheckErrorKind::ValuesMissingForUdtFields { ref field_names },
-                ) = err.kind
-                else {
-                    panic!("unexpected error kind: {:?}", err.kind)
-                };
-                assert_eq!(field_names, &["c"]);
-            }
-
-            // UDT fields incompatible types - field type check failed
-            {
-                let typ = udt_def_with_fields([
-                    ("a", ColumnType::Native(NativeType::Blob)),
-                    ("b", ColumnType::Native(NativeType::Int)),
-                ]);
-                let err = Udt::type_check(&typ).unwrap_err();
-                let err = get_typeck_err_inner(&err);
-                assert_eq!(err.rust_name, std::any::type_name::<Udt>());
-                assert_eq!(err.cql_type, typ);
-                let BuiltinTypeCheckErrorKind::UdtError(
-                    UdtTypeCheckErrorKind::FieldTypeCheckFailed {
-                        ref field_name,
-                        ref err,
-                    },
-                ) = err.kind
-                else {
-                    panic!("unexpected error kind: {:?}", err.kind)
-                };
-                assert_eq!(field_name.as_str(), "a");
-                let err = get_typeck_err_inner(err);
-                assert_eq!(err.rust_name, std::any::type_name::<&str>());
-                assert_eq!(err.cql_type, ColumnType::Native(NativeType::Blob));
-                assert_matches!(
-                    err.kind,
-                    BuiltinTypeCheckErrorKind::MismatchedType {
-                        expected: &[
-                            ColumnType::Native(NativeType::Ascii),
-                            ColumnType::Native(NativeType::Text)
-                        ]
-                    }
-                );
-            }
-        }
-
-        // Deserialization errors
-        {
-            // Got null
-            {
-                let typ = udt_def_with_fields([
-                    ("c", ColumnType::Native(NativeType::Boolean)),
-                    ("a", ColumnType::Native(NativeType::Blob)),
-                    ("b", ColumnType::Native(NativeType::Int)),
-                ]);
-
-                let err = Udt::deserialize(&typ, None).unwrap_err();
-                let err = get_deser_err_inner(&err);
-                assert_eq!(err.rust_name, std::any::type_name::<Udt>());
-                assert_eq!(err.cql_type, typ);
-                assert_matches!(err.kind, BuiltinDeserializationErrorKind::ExpectedNonNull);
-            }
-
-            // UDT field deserialization failed
-            {
-                let typ = udt_def_with_fields([
-                    ("a", ColumnType::Native(NativeType::Ascii)),
-                    ("c", ColumnType::Native(NativeType::Boolean)),
-                ]);
-
-                let udt_bytes = UdtSerializer::new()
-                    .field("alamakota".as_bytes())
-                    .field(&42_i16.to_be_bytes())
-                    .finalize();
-
-                let err = deserialize::<Udt>(&typ, &udt_bytes).unwrap_err();
-
-                let err = get_deser_err(&err);
-                assert_eq!(err.rust_name, std::any::type_name::<Udt>());
-                assert_eq!(err.cql_type, typ);
-                let BuiltinDeserializationErrorKind::UdtError(
-                    UdtDeserializationErrorKind::FieldDeserializationFailed {
-                        ref field_name,
-                        ref err,
-                    },
-                ) = err.kind
-                else {
-                    panic!("unexpected error kind: {:?}", err.kind)
-                };
-                assert_eq!(field_name.as_str(), "c");
-                let err = get_deser_err_inner(err);
-                assert_eq!(err.rust_name, std::any::type_name::<bool>());
-                assert_eq!(err.cql_type, ColumnType::Native(NativeType::Boolean));
-                assert_matches!(
-                    err.kind,
-                    BuiltinDeserializationErrorKind::ByteLengthMismatch {
-                        expected: 1,
-                        got: 2,
-                    }
-                );
-            }
-        }
-    }
-
-    // Strict ordering
-    {
-        #[derive(scylla_macros::DeserializeValue, PartialEq, Eq, Debug)]
-        #[scylla(crate = "crate", flavor = "enforce_order", forbid_excess_udt_fields)]
-        struct Udt<'a> {
-            a: &'a str,
-            #[scylla(skip)]
-            x: String,
-            b: Option<i32>,
-            #[scylla(allow_missing)]
-            c: bool,
-        }
-
-        // Type check errors
-        {
-            // Not UDT
-            {
-                let typ = ColumnType::Collection {
-                    frozen: false,
-                    typ: CollectionType::Map(
-                        Box::new(ColumnType::Native(NativeType::Ascii)),
-                        Box::new(ColumnType::Native(NativeType::Blob)),
-                    ),
-                };
-                let err = Udt::type_check(&typ).unwrap_err();
-                let err = get_typeck_err_inner(&err);
-                assert_eq!(err.rust_name, std::any::type_name::<Udt>());
-                assert_eq!(err.cql_type, typ);
-                let BuiltinTypeCheckErrorKind::UdtError(UdtTypeCheckErrorKind::NotUdt) = err.kind
-                else {
-                    panic!("unexpected error kind: {:?}", err.kind)
-                };
-            }
-
-            // UDT too few fields
-            {
-                let typ = udt_def_with_fields([("a", ColumnType::Native(NativeType::Text))]);
-                let err = Udt::type_check(&typ).unwrap_err();
-                let err = get_typeck_err_inner(&err);
-                assert_eq!(err.rust_name, std::any::type_name::<Udt>());
-                assert_eq!(err.cql_type, typ);
-                let BuiltinTypeCheckErrorKind::UdtError(UdtTypeCheckErrorKind::TooFewFields {
-                    ref required_fields,
-                    ref present_fields,
-                }) = err.kind
-                else {
-                    panic!("unexpected error kind: {:?}", err.kind)
-                };
-                assert_eq!(required_fields.as_slice(), &["a", "b"]);
-                assert_eq!(present_fields.as_slice(), &["a".to_string()]);
-            }
-
-            // excess fields in UDT
-            {
-                let typ = udt_def_with_fields([
-                    ("a", ColumnType::Native(NativeType::Text)),
-                    ("b", ColumnType::Native(NativeType::Int)),
-                    ("d", ColumnType::Native(NativeType::Boolean)),
-                ]);
-                let err = Udt::type_check(&typ).unwrap_err();
-                let err = get_typeck_err_inner(&err);
-                assert_eq!(err.rust_name, std::any::type_name::<Udt>());
-                assert_eq!(err.cql_type, typ);
-                let BuiltinTypeCheckErrorKind::UdtError(UdtTypeCheckErrorKind::ExcessFieldInUdt {
-                    ref db_field_name,
-                }) = err.kind
-                else {
-                    panic!("unexpected error kind: {:?}", err.kind)
-                };
-                assert_eq!(db_field_name.as_str(), "d");
-            }
-
-            // UDT fields switched - field name mismatch
-            {
-                let typ = udt_def_with_fields([
-                    ("b", ColumnType::Native(NativeType::Int)),
-                    ("a", ColumnType::Native(NativeType::Text)),
-                ]);
-                let err = Udt::type_check(&typ).unwrap_err();
-                let err = get_typeck_err_inner(&err);
-                assert_eq!(err.rust_name, std::any::type_name::<Udt>());
-                assert_eq!(err.cql_type, typ);
-                let BuiltinTypeCheckErrorKind::UdtError(UdtTypeCheckErrorKind::FieldNameMismatch {
-                    position,
-                    ref rust_field_name,
-                    ref db_field_name,
-                }) = err.kind
-                else {
-                    panic!("unexpected error kind: {:?}", err.kind)
-                };
-                assert_eq!(position, 0);
-                assert_eq!(rust_field_name.as_str(), "a".to_owned());
-                assert_eq!(db_field_name.as_str(), "b".to_owned());
-            }
-
-            // UDT fields incompatible types - field type check failed
-            {
-                let typ = udt_def_with_fields([
-                    ("a", ColumnType::Native(NativeType::Blob)),
-                    ("b", ColumnType::Native(NativeType::Int)),
-                ]);
-                let err = Udt::type_check(&typ).unwrap_err();
-                let err = get_typeck_err_inner(&err);
-                assert_eq!(err.rust_name, std::any::type_name::<Udt>());
-                assert_eq!(err.cql_type, typ);
-                let BuiltinTypeCheckErrorKind::UdtError(
-                    UdtTypeCheckErrorKind::FieldTypeCheckFailed {
-                        ref field_name,
-                        ref err,
-                    },
-                ) = err.kind
-                else {
-                    panic!("unexpected error kind: {:?}", err.kind)
-                };
-                assert_eq!(field_name.as_str(), "a");
-                let err = get_typeck_err_inner(err);
-                assert_eq!(err.rust_name, std::any::type_name::<&str>());
-                assert_eq!(err.cql_type, ColumnType::Native(NativeType::Blob));
-                assert_matches!(
-                    err.kind,
-                    BuiltinTypeCheckErrorKind::MismatchedType {
-                        expected: &[
-                            ColumnType::Native(NativeType::Ascii),
-                            ColumnType::Native(NativeType::Text)
-                        ]
-                    }
-                );
-            }
-        }
-
-        // Deserialization errors
-        {
-            // Got null
-            {
-                let typ = udt_def_with_fields([
-                    ("a", ColumnType::Native(NativeType::Text)),
-                    ("b", ColumnType::Native(NativeType::Int)),
-                    ("c", ColumnType::Native(NativeType::Boolean)),
-                ]);
-
-                let err = Udt::deserialize(&typ, None).unwrap_err();
-                let err = get_deser_err_inner(&err);
-                assert_eq!(err.rust_name, std::any::type_name::<Udt>());
-                assert_eq!(err.cql_type, typ);
-                assert_matches!(err.kind, BuiltinDeserializationErrorKind::ExpectedNonNull);
-            }
-
-            // Bad field format
-            {
-                let typ = udt_def_with_fields([
-                    ("a", ColumnType::Native(NativeType::Text)),
-                    ("b", ColumnType::Native(NativeType::Int)),
-                    ("c", ColumnType::Native(NativeType::Boolean)),
-                ]);
-
-                let udt_bytes = UdtSerializer::new()
-                    .field(b"alamakota")
-                    .field(&42_i64.to_be_bytes())
-                    .field(&[true as u8])
-                    .finalize();
-
-                let udt_bytes_too_short = udt_bytes.slice(..udt_bytes.len() - 1);
-                assert!(udt_bytes.len() > udt_bytes_too_short.len());
-
-                let err = deserialize::<Udt>(&typ, &udt_bytes_too_short).unwrap_err();
-
-                let err = get_deser_err(&err);
-                assert_eq!(err.rust_name, std::any::type_name::<Udt>());
-                assert_eq!(err.cql_type, typ);
-                let BuiltinDeserializationErrorKind::RawCqlBytesReadError(_) = err.kind else {
-                    panic!("unexpected error kind: {:?}", err.kind)
-                };
-            }
-
-            // UDT field deserialization failed
-            {
-                let typ = udt_def_with_fields([
-                    ("a", ColumnType::Native(NativeType::Text)),
-                    ("b", ColumnType::Native(NativeType::Int)),
-                    ("c", ColumnType::Native(NativeType::Boolean)),
-                ]);
-
-                let udt_bytes = UdtSerializer::new()
-                    .field(b"alamakota")
-                    .field(&42_i64.to_be_bytes())
-                    .field(&[true as u8])
-                    .finalize();
-
-                let err = deserialize::<Udt>(&typ, &udt_bytes).unwrap_err();
-
-                let err = get_deser_err(&err);
-                assert_eq!(err.rust_name, std::any::type_name::<Udt>());
-                assert_eq!(err.cql_type, typ);
-                let BuiltinDeserializationErrorKind::UdtError(
-                    UdtDeserializationErrorKind::FieldDeserializationFailed {
-                        ref field_name,
-                        ref err,
-                    },
-                ) = err.kind
-                else {
-                    panic!("unexpected error kind: {:?}", err.kind)
-                };
-                assert_eq!(field_name.as_str(), "b");
-                let err = get_deser_err_inner(err);
-                assert_eq!(err.rust_name, std::any::type_name::<Option<i32>>());
-                assert_eq!(err.cql_type, ColumnType::Native(NativeType::Int));
-                assert_matches!(
-                    err.kind,
-                    BuiltinDeserializationErrorKind::ByteLengthMismatch {
-                        expected: 4,
-                        got: 8,
-                    }
-                );
-            }
-        }
-    }
-}
-
-#[test]
 fn metadata_does_not_bound_deserialized_values() {
     /* It's important to understand what is a _deserialized value_. It's not just
      * an implementor of DeserializeValue; there are some implementors of DeserializeValue
@@ -3218,28 +2349,6 @@ fn metadata_does_not_bound_deserialized_values() {
         };
         let decoded_map_str_int_res = deserialize::<HashMap<&str, i32>>(&map_typ, &bytes);
 
-        // UDT
-        let udt_typ = ColumnType::UserDefinedType {
-            frozen: false,
-            definition: Arc::new(UserDefinedType {
-                name: "udt".into(),
-                keyspace: "ks".into(),
-                field_types: vec![
-                    ("bytes".into(), ColumnType::Native(NativeType::Blob)),
-                    ("text".into(), ColumnType::Native(NativeType::Text)),
-                ],
-            }),
-        };
-        #[derive(DeserializeValue)]
-        #[scylla(crate=crate)]
-        struct Udt<'frame> {
-            #[expect(dead_code)]
-            bytes: &'frame [u8],
-            #[expect(dead_code)]
-            text: &'frame str,
-        }
-        let decoded_udt_res = deserialize::<Udt>(&udt_typ, &bytes);
-
         (
             decoded_blob_res,
             decoded_str_res,
@@ -3248,7 +2357,6 @@ fn metadata_does_not_bound_deserialized_values() {
             decoded_set_str_res,
             decoded_set_string_res,
             decoded_map_str_int_res,
-            decoded_udt_res,
         )
     };
 }
